CHROMIUM: board-seaboard: mxt: Fix orientation for seaboard and asymptote
The seaboard and asymptote touchscreens use the Atmel mxt touch controller.
These touchpads are rotated 90 degrees relative to the Atmel default
configuration.  Use the board configuration file to set this orientation.

Note: previously the X axis was inverted.  This inversion was corrected
using a calibration in the xf86-input-evdev .conf file.  Thus, this patch
must land with a corresponding patch to the asymptote/seaboard .conf file.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:20369
TEST=(1) Use evtest/mtplot
  X-coordinate should increase from left to right
  Y-coordinate should increase from top to bottom
     (2) Use touchui Chromium (with evdev .conf calibration fix:
                               gerrit.chromium.org/gerrit/7588)
  touch elements should still be clickable

Change-Id: I2c31280e944ad2831ecde32d41063561ff1b76f6
Reviewed-on: http://gerrit.chromium.org/gerrit/7585
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>