CHROMIUM: tegra: Make earlyprintk choose a UART at runtime.
With this change we automatically detect which UART to use for
earlyprintk (and for printing during decompression).  The
detection involves coordination with the bootloader: it's expected
that the bootloader will leave a 'D' in the UART scratchpad
register for whatever UART we should use for debugging.

If we don't find any UARTs that were specified, we'll fall back to
the UART that was specified during config time:
CONFIG_TEGRA_DEBUG_UART_XXX.

As a side effect of this change, earlyprintk will no longer fail
if you've specified CONFIG_TEGRA_DEBUG_UART_NONE.

BUG=chromium-os:19656
TEST=Manual
1. Applied future change to set CONFIG_TEGRA_DEBUG_UART_NONE.
2. Applied future change to u-boot set a 'D' in the scratchpad of
   the right UART.
3. Installed legacy u-boot and made sure I had earlyprintk by doing
   setenv dev_extras earlyprintk
   saveenv
4. Booted Searboard.  Saw the "Uncompressing Linux..." message as
   well as early printouts.
5. Rebooted, holding space on serial console while booting to drop
   into u-boot.
6. Manually unset the scratchpad register on UART D:
   md.l 0x7000631c 1
   mw.l 0x7000631c ' '
   md.l 0x7000631c 1
7. Booted:
   boot
8. Saw no early printouts.

I also tried changing the default kernel config to UARTD.  Then and
repeated steps #1-#7.  In this case, in step #8 I still saw all
of the messages on the console.

Change-Id: Icc3204022dc440f78ff796c0a59f48884ab1cca4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8148
Reviewed-by: Stephen Warren <swarren@nvidia.com>