ath6kl: Replace spin_lock_irqsave with spin_lock_bh
It is not necessary to use spinlock primitive to
protect data which is accessed in hard irq context as
nothing is running in hard irq with this driver. The
spinlock primitive meant to protect data in softirq
context is more appropriate.

Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@qca.qualcomm.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
(cherry picked from commit 5999264f4bee79f69bd62cab275a09563d195ceb)

BUG=chrome-os-partner:5925
TEST=build kernel

Change-Id: I2a06fed4de9501ae257f715d42a994d103daf315