ath6kl: Avoid processing failed rx packets
It is not necessary to process an htc_packet which is allocated for rx
but failed in sdio rx. Though it does not fix any real issue, it does
not make much sense to process the failed frame.

Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@qca.qualcomm.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
(cherry picked from commit 76d0d384d9f4492423ea7557de836ad4fa9f1c8a)

BUG=chrome-os-partner:5925
TEST=matfunc+secmat+manager+perf+roaming for regression

Change-Id: Ib80e375034c0072add9a1670d909baa016a0299f