UPSTREAM: CHROMIUM: driver: Google Memory Console
This patch introduces the 'memconsole' driver.

Our firmware gives us access to an in-memory log of the firmware's
output.   This gives us visibility in a data-center of headless machines
as to what the firmware is doing.

The memory console is found by the driver by finding a header block in
the EBDA.  The buffer is then copied out, and is exported to userland in
the file /sys/firmware/log.

Signed-off-by: San Mehat <san@google.com>
Signed-off-by: Mike Waychison <mikew@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I0b04635c4e374459103b0a305eb68b5f4066222c
Reviewed-on: http://gerrit.chromium.org/gerrit/8779
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>