CHROMIUM: HACK: ARM: tegra: asymptote: Temporarily set carveout to 128M
Current Asymptotes have older firmware which will not boot with the 256MB
carveout. This patch will set the carveout to 128MB for Asymptotes until we can
flash all of the devices. Once all Asymptotes are up-to-date, I will remove this
they will use the 256MB carveout like the other tegra targets.

BUG=None
TEST=Built and ran on Asymptote, ensured the system booted.

Change-Id: I018e62825aad5476a20ea7cf021085f0b32a5061
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/9891