Input: cyapa - SMBus support
Support for Cypress APA trackpads using
the SMBus protocol.

The i2c/smbus adapter can be queried for functionality,
and we can select the appropriate type.

In the case where both are supported by the interface, the
driver will prefer i2c. A future change may detect failures in i2c
mode and switch to smbus if both are supported by the adapter in
cyapa_get_trackpad_run_mode.

BUG=chrome-os-partner:5698
TEST=Test on i2c cyapa pad. Make sure no functional
difference after this change.
Test on Smbus cyapa pad on a Smbus system. Make sure
that the pad is functional in both cold boot and
warm boot scenarios.

Change-Id: I62d133ab82cb0db037004f62cd8c787a8e2f6006
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/10560
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>