CHROMIUM: gobi: qc_deregister in qcnet_disconnect, not _unbind
Right now, if qc_register fails, we are obligated to call
usbnet_disconnect to undo usbnet_probe, but by calling that, we call
qc_deregister, which tries to undo what qc_register already undid.

So, move the call to qc_deregister to qcnet_disconnect.  This makes it
symmetric with the call to qc_register in qcnet_bind, and allows us to
recover after a failed qc_register in qcnet_probe.

BUG=chromium-os:22320
TEST=Adhoc; forcing qmidms_getmeid to fail no longer panics

Signed-off-by: ttuttle <ttuttle@chromium.org>
Change-Id: I78d3415c73a4a3526ef022abe0ff82db3623e712
Reviewed-on: https://gerrit.chromium.org/gerrit/10935
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>